+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - prompt to install SDK on startup if not already installed
 - validate the specified SDK path
 - integrate launch into the existing dartlang plugin launching framework
+- update 'Create new Project' to use 'dartino create project' command
 
 ## 0.0.5
 - hide atom-toolbar on startup (until Dartino launch is integrated in dartlang)